backend/european_champonship_kettlebell/requirements.txt
asgiref==3.8.1
cfgv==3.4.0
diff-match-patch==20241021
distlib==0.3.9
Django==5.2
django-import-export==4.3.7
django-jazzmin==3.0.1
djangorestframework==3.16.0
factory_boy==3.3.3
Faker==37.1.0
filelock==3.18.0
gunicorn==23.0.0
identify==2.6.9
isort==6.0.1
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.7
pre_commit==4.2.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
PyYAML==6.0.2
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
virtualenv==20.29.3
